requirements.txt
autopep8
attrs
black
cattrs
flake8
isort
mypy
pycodestyle
pytest
python-dotenv
